feat: restructure agents extension; plan-persona write scope; orchestrator auto-pick model
Three coordinated changes to the agents extension.

1. Folder restructure under src/extensions/agents/

   25 files reorganized into manager/ personas/ resolution/ prompt/
   memory/ ui/ subfolders. Old flat paths kept as re-export shims so
   nothing outside agents/ has to change. ui/ already existed and stayed.

2. Plan persona — write scope under .kimchi/plans/

   - plan persona gains write+edit tools (was read-only).
   - persona body instructs the agent to only write under .kimchi/plans/.
   - manager/agent-runner.ts sets process.env.KIMCHI_AGENT_PERSONA on
     spawn so the permissions hook can scope writes per-persona.
   - permissions/index.ts: when KIMCHI_AGENT_PERSONA=plan, write/edit
     calls outside .kimchi/plans/ are blocked with a notify; calls
     inside the dir pass through normal rules.

3. Orchestrator auto-pick fallback for Agent tool

   New helper: orchestration/model-registry/recommend.ts exports
   recommendModel({ strengths, needsVision?, preferTier? }) which
   filters MODEL_CAPABILITIES and applies tier fallback (light →
   standard → heavy / etc). Deterministic order.

   resolution/invocation-config.ts new fallback chain for the model:
     1. params.model (LLM picked)              wins
     2. agentConfig.models[0]                   stable persona default
     3. recommendModel(agentConfig.strengths)   orchestrator picks for persona
     4. recommendModel([currentPhase])          orchestrator picks for phase
     5. inherit parent

   personas/types.ts gains optional strengths[] and preferTier fields;
   default-agents.ts populates them for personas with a clear strength
   profile (explore→[explore], researcher→[research], plan→[plan],
   worker→[build], reviewer→[review]). general-purpose leaves them
   undefined to inherit parent.

   The orchestrator is unchanged in role: still informs the parent's
   system prompt about model strengths so the parent picks. Auto-pick
   only kicks in when the parent explicitly omits model on the Agent
   call AND the persona has strengths declared. Parent always wins
   when it specifies model.
