+++ src/extensions/orchestration/model-registry/index.ts
@@ -4,3 +4,5 @@ export type { ModelRegistryWarning } from "./model-registry.js"
 export type { ModelStrength, ModelTier, ModelCapabilities, OrchestrationModelDescriptor } from "./types.js"
 export { modelsForStrength, modelsForAnyStrength } from "./by-strength.js"
 export type { StrengthResolveOptions } from "./by-strength.js"
+export { recommendModel } from "./recommend.js"
+export type { RecommendOptions, RecommendResult } from "./recommend.js"
