+++ src/extensions/permissions/index.ts
@@ -1,3 +1,4 @@
+import { resolve } from "node:path"
 import type { Api, Model } from "@mariozechner/pi-ai"
 import type { ExtensionAPI, ExtensionContext, ToolCallEvent } from "@mariozechner/pi-coding-agent"
 import { isKeyRelease, matchesKey } from "@mariozechner/pi-tui"
@@ -13,6 +14,21 @@ import { evaluateRules, parseRules, stringifyRule } from "./rules.js"
 import { SessionMemory } from "./session-memory.js"
 import { isReadOnlyBashCommand, isReadOnlyTool } from "./taxonomy.js"
 import { BUILTIN_DENY, DEFAULT_CONFIG, type PermissionMode, type Rule } from "./types.js"
+
+/**
+ * Check whether a file path is within .kimchi/plans/ relative to cwd.
+ * Accepts both relative paths (starting with .kimchi/plans/) and
+ * absolute paths under <cwd>/.kimchi/plans/.
+ * Path traversal is prevented by resolving to an absolute path first.
+ */
+export function isWithinKimchiPlans(filePath: string, cwd: string): boolean {
+	const normalizedCwd = cwd.endsWith("/") ? cwd : `${cwd}/`
+	const plansDir = `${normalizedCwd}.kimchi/plans/`
+
+	// Resolve to absolute, normalizing any ".." components
+	const abs = resolve(cwd, filePath)
+	return abs.startsWith(plansDir)
+}
 
 /**
  * DANGER: Bypass flag that disables ALL permission checks.
@@ -271,6 +287,23 @@ export default function permissionsExtension(pi: ExtensionAPI): void {
 	pi.on("tool_call", async (event, ctx) => {
 		const toolName = event.toolName.toLowerCase()
 		const input = event.input as Record<string, unknown>
+
+		// Plan persona path-scope enforcement: when KIMCHI_AGENT_PERSONA=plan (case-insensitive),
+		// write and edit are only allowed for .kimchi/plans/* paths.
+		if (process.env.KIMCHI_AGENT_PERSONA?.toLowerCase() === "plan") {
+			if (toolName === "write" || toolName === "edit") {
+				const filePath =
+					typeof input.file_path === "string" ? input.file_path : typeof input.path === "string" ? input.path : ""
+				if (filePath && !isWithinKimchiPlans(filePath, ctx.cwd)) {
+					return {
+						block: true,
+						reason: `Plan persona: ${toolName} is restricted to .kimchi/plans/ files. The path "${filePath}" is outside that scope.`,
+					}
+				}
+				// path is within .kimchi/plans/ — allow without further checks
+				return undefined
+			}
+		}
 
 		// Re-evaluation loop: when a permission prompt is dismissed because the user
 		// changed mode via shift+tab, we re-evaluate the tool call under the new mode.
